Add files via upload
The code for the 2048 game. 4x4 version, the behavior is logged in the txt file.
It is a replica of the 2048 game produced by Italian developer Gabriel Cirulli. Detailed driving may vary, and the GUI has been implemented using Qt.
In particular, the 2048 game is also on his Github.
https://github.com/gabrielecirulli/2048
It's just a simulation of this as a task, and there's no revenue generation.